refactor: extract QueryStatistics class into its own module
