+++ src/presentation/table-presenter.ts
@@ -1,4 +1,4 @@
-import { QueryStatistics } from "../mesh/mesh-adapter.ts";
+import { QueryStatistics } from "../mesh/query-statistics.ts";
 import { MeshTable, TableGenerator } from "./mesh-table.ts";
 import { Presenter } from "./presenter.ts";
